feat: update incedent details component
