+++ src/components/IncidentDetails/data.js
@@ -1,0 +1,20 @@
+export const personRespondents = [
+  {
+    label: "Lindsay Walton",
+    value: "Lindsay Walton",
+    avatar:
+      "https://images.unsplash.com/photo-1517841905240-472988babdf9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=3&w=256&h=256&q=80"
+  },
+  {
+    label: "Eduardo Benz",
+    value: "Eduardo Benz",
+    avatar:
+      "https://images.unsplash.com/photo-1520785643438-5bf77931f493?ixlib=rb-=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=8&w=256&h=256&q=80"
+  },
+  {
+    label: "Test Person",
+    value: "2e8872ff-eaa1-473f-9fd9-823235804d30",
+    avatar:
+      "https://images.unsplash.com/photo-1550525811-e5869dd03032?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
+  }
+];
