+++ src/index.css
@@ -13,10 +13,10 @@
   }
 
   .btn-secondary {
-    @apply inline-flex items-center justify-center border border-gray-300 shadow-sm font-medium rounded-md text-indigo-500 bg-indigo-100 hover:bg-indigo-200 focus:ring-offset-white focus:ring-indigo-500 focus:outline-none focus:ring-2 focus:ring-offset-2;
+    @apply inline-flex items-center justify-center border-none shadow-sm font-medium rounded-md text-indigo-500 bg-indigo-100 hover:bg-indigo-200 focus:ring-offset-white focus:ring-indigo-500 focus:outline-none focus:ring-2 focus:ring-offset-2;
   }
   .btn-secondary-disabled {
-    @apply inline-flex items-center justify-center border border-gray-300 shadow-sm font-medium rounded-md text-indigo-100 bg-gray-100 opacity-50 cursor-not-allowed
+    @apply inline-flex items-center justify-center border-none shadow-sm font-medium rounded-md text-indigo-100 bg-gray-100 opacity-50 cursor-not-allowed
   }
   .btn-secondary-xs {
     @apply px-2.5 py-1.5 rounded-4px text-xs leading-4 font-medium
